requirements.txt
fastapi
uvicorn
pymongo[srv]
python-dotenv
pydantic
transformers
torch
torchvision
accelerate
pillow
zai-sdk
sniffio
langchain
langchain-huggingface
bcrypt
pydantic[email]